Launch python notebook in a subprocess
